requirements.txt
chromadb
langchain
python-dotenv
pypdf
openpyxl
python-docx